Code review updates. Add underscore to private methods, rename options with transaction_options, split ReadWrite class into ReadWrite and ReadOnly
